refactor(AppState): split the 2308-line class across six files
AppState is the app's central state object and had grown to 2308 lines
spanning eight unrelated concerns. Moved the behaviour into extensions
in sibling files, keeping AppState.swift for stored properties (Swift
extensions cannot hold them) and the Defaults-backed settings:

    AppState.swift                    2308 -> 805
    AppState+BackendSync.swift               639
    AppState+Conflicts.swift                 372
    AppState+LiveActivity.swift              141
    AppState+RevisionPolling.swift           127
    AppState+PushServer.swift                113
    AppState+Account.swift                   107

No logic moved between phases — every function is byte-identical to
where it was, only relocated.

The cost worth naming: `private` in Swift is file-scoped, so 19 stored
properties the extensions touch had to widen to internal. `lastSyncSource`
did not — roughly a hundred view files read it and none should be able
to assign it, so it keeps `private(set)` and BackendSync goes through a
new `recordSyncSource(_:)` instead of being demoted to a plain `var`.

Also fixes the test target, which turned out not to compile at all.
TigerDuckTests references WidgetTimelineDerivation via
`@testable import TigerDuck`, but that file lives in the widgets
extension group and was never a member of the app target — 26 "cannot
find in scope" errors, so the whole bundle has been unbuildable. Added
it (and WidgetTaipei, which it needs) to the app target's membership
exception set, the same way WidgetSnapshotStore already was.

check_macos_sources.py now parses those exception sets out of
project.pbxproj rather than hardcoding the one file it knew about — a
hardcoded list goes stale exactly the way the allowlist does, which is
the failure this script exists to catch.
