+++ tools/macos-excluded-sources.txt
@@ -179,3 +179,10 @@ FlipDetector.swift
 
 # Shared — used only by the iOS-side surfaces above
 TaipeiCalendar.swift
+
+# App — AppState extensions that are entirely iOS-only
+# ActivityKit has no macOS counterpart; the whole file body sits inside
+# #if os(iOS), so compiling it on the Mac would only produce an empty
+# extension. Listed here rather than allowlisted so that adding a
+# cross-platform member to it is a decision someone has to make.
+AppState+LiveActivity.swift
